+++ src/main/java/de/samply/email/EmailRecipient.java
@@ -4,5 +4,5 @@ import de.samply.user.roles.ProjectRole;
 
 import java.util.Optional;
 
-public record EmailRecipient(String email, Optional<String> bridgehead, ProjectRole role) {
+public record EmailRecipient(String email, Optional<String> projectCode, Optional<String> bridgehead, ProjectRole role) {
 }